Move MockDataStream tests to a separate file and reduce duplication
